Add support for TRIM on raw files on macOS
The macOS API `fcntl(F_PUNCHHOLE)` will successfully punch holes
in files on APFS and free the backing storage so we can use this
to implement the host side of TRIM.

The arguments to the `fcntl` need to be aligned to the filesystems
`f_bsize` as returned by `statfs` which on my Mac's SSD is 4096.

Unfortunately exposing the 4096 byte sectors to the guest seems to
trigger a bug elsewhere in the ACHI implementation so this patch
maintains a virtual sector size of 512 bytes. This means that the
guest will send 512 byte-aligned requests, we must round these to
(eg) 4096 byte-aligned requests and manually zero any leading and
trailing sectors. For the zeroing we allocate a `delete_zero_buf`
per block device and pre-zero it.

When the device is opened we perform a test `F_PUNCHHOLE` of length
0 at offset 0 and check the return value. If the call succeeds
then we assume sparse files is supported (e.g. APFS) and if the
call fails we assume sparse files are not supported (e.g. HFS+)

Signed-off-by: David Scott <dave.scott@docker.com>