virtio-blk: only advertise DISCARD if the backing file supports it
We already have an API `blockif_candelete` which tells us whether the
backing file supports trim/discard or not. We already use this in
the AHCI implementation to decide whether to expose the capability
bits.

This patch clears the `DISCARD` feature bit in the negotiation if
we discover that the backing file does not support it (e.g. if it
is a raw file on HFS+). Previously all the feature bits in virtio-blk
were global constants.

Signed-off-by: David Scott <dave@recoil.org>